requirements.txt
django==3.2.25
django-cors-headers==3.10.1
djangorestframework==3.15.1
pymysql
django-simpleui==2025.1.13
python-dotenv
django-apscheduler==0.6.2
pure-sasl==0.6.2
thrift-sasl==0.4.3
pyhive==0.7.0
pyhdfs==0.3.1
loguru
pytz_deprecation_shim

